third_party/blink/web_tests/external/wpt/css/css-fonts/math-script-level-and-math-style/math-script-level-002.tentative-expected.txt

This is a testharness.js-based test.
FAIL auto assert_equals: expected 355 but got 500
FAIL auto ; starting from level 7 assert_equals: expected 355 but got 500
FAIL add(<integer>) assert_equals: expected 100 but got 71
FAIL add(<integer>) ; starting from level 3 assert_equals: expected 100 but got 71
FAIL <integer> assert_equals: expected 100 but got 71
FAIL <integer> ; starting from level 50 assert_equals: expected 100 but got 71
Harness: the test ran to completion.
